+++ src/utils/pixel.rs
@@ -4,12 +4,12 @@ use std::{
 };
 
 // values are defined in a range [0.0, 1.0]
-#[derive(PartialEq, Debug, Clone, Copy)]
+#[derive(PartialEq, Debug, Clone, Copy, Default)]
 pub struct RGB {
-    pub r: f64,
-    pub g: f64,
-    pub b: f64,
-    pub a: f64,
+    pub r: f32,
+    pub g: f32,
+    pub b: f32,
+    pub a: f32,
 }
 
 impl Add for RGB {
@@ -43,10 +43,10 @@ impl Sub for RGB {
 impl RGB {
     pub fn from_u8(r: u8, g: u8, b: u8, a: u8) -> RGB {
         RGB {
-            r: r as f64 / 255.0,
-            g: g as f64 / 255.0,
-            b: b as f64 / 255.0,
-            a: a as f64 / 255.0,
+            r: r as f32 / 255.0,
+            g: g as f32 / 255.0,
+            b: b as f32 / 255.0,
+            a: a as f32 / 255.0,
         }
     }
 
@@ -56,9 +56,9 @@ impl RGB {
         let g_str = &clean_string[2..4];
         let b_str = &clean_string[4..6];
 
-        let r = u32::from_str_radix(r_str, 16)? as f64 / 255.0;
-        let g = u32::from_str_radix(g_str, 16)? as f64 / 255.0;
-        let b = u32::from_str_radix(b_str, 16)? as f64 / 255.0;
+        let r = u32::from_str_radix(r_str, 16)? as f32 / 255.0;
+        let g = u32::from_str_radix(g_str, 16)? as f32 / 255.0;
+        let b = u32::from_str_radix(b_str, 16)? as f32 / 255.0;
 
         Ok(RGB { r, g, b, a: 1.0 })
     }
@@ -79,7 +79,7 @@ impl RGB {
         r << 16 | g << 8 | b
     }
 
-    pub fn grayscale(&self) -> f64 {
+    pub fn grayscale(&self) -> f32 {
         0.299 * self.r + 0.587 * self.g + 0.114 * self.b
     }
 
@@ -93,13 +93,13 @@ impl RGB {
         }
     }
 
-    pub fn add_luminosity(&mut self, amount: f64) {
+    pub fn add_luminosity(&mut self, amount: f32) {
         self.r += amount;
         self.g += amount;
         self.b += amount;
     }
 
-    pub fn set_value(&mut self, value: f64) {
+    pub fn set_value(&mut self, value: f32) {
         self.r = (value).clamp(0.0, 1.0);
         self.g = (value).clamp(0.0, 1.0);
         self.b = (value).clamp(0.0, 1.0);
